+++ modules/conversion-center/traits/conversion-center-controls-trait.php
@@ -3,6 +3,8 @@
 namespace Elementor\Modules\ConversionCenter\Traits;
 
 use Elementor\Controls_Manager;
+use Elementor\Shapes;
+use Elementor\Utils;
 
 trait Conversion_Center_Controls_Trait {
 
@@ -172,5 +174,48 @@ trait Conversion_Center_Controls_Trait {
 			$prefix . '_border_color',
 			array_merge( $border_color, $border_color_args )
 		);
+	}
+
+	protected function get_shape_divider( $side = 'bottom' ) {
+		$settings = $this->settings;
+		$base_setting_key = "identity_section_style_cover_divider_$side";
+		$file_name = $settings[ $base_setting_key ];
+
+		if ( empty( $file_name ) ) {
+			return [];
+		}
+
+		$negative = ! empty( $settings[ $base_setting_key . '_negative' ] );
+		$shape_path = Shapes::get_shape_path( $file_name, $negative );
+
+		if ( ! is_file( $shape_path ) || ! is_readable( $shape_path ) ) {
+			return [];
+		}
+
+		return [
+			'negative' => $negative,
+			'svg' => Utils::file_get_contents( $shape_path ),
+		];
+	}
+
+	protected function print_shape_divider( $side = 'bottom' ) {
+		$shape_divider = $this->get_shape_divider( $side );
+
+		if ( empty( $shape_divider ) ) {
+			return;
+		}
+		?>
+		<div
+			class="elementor-shape elementor-shape-<?php echo esc_attr( $side ); ?>"
+			data-negative="<?php
+			echo esc_attr( $shape_divider['negative'] ? 'true' : 'false' );
+			?>"
+		>
+			<?php
+			// PHPCS - The file content is being read from a strict file path structure.
+			echo $shape_divider['svg']; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+			?>
+		</div>
+		<?php
 	}
 }
